Support for URL based API versioning. First version is pre1.
API version is part of the URL specification. The unversioned API url is
still provided, the default version for that is configured in project
settings.

Version checks in code are done like `if request.version == 'pre1'`. The
intention is that breaking changes are coded as the default path and old
functionality is provided with conditionals, deprecated and removed quickly
in the release cycle. Deprecated code should be clearly commented as
follows: `# DEPRECATED, used by mobile v0.2.0`.